refactor(element): decompose IC into ICLoader + ICRenderer + ICSimulation
IC.h/IC.cpp (198/1070 lines) mixed sub-circuit loading, rendering, and
internal simulation directly. Split into three static-method friend
collaborators, mirroring the GraphicElement/GraphicElementSerializer
split — but since IC is already heavily friended by tests (TestArduino,
TestIC, TestDanglingPointer reach internal element/port vectors and call
resetInternalState() directly), all state stays on IC itself; the three
collaborators are pure logic extractions, not new owner-back-pointer
classes with their own state.

ICLoader: loadFile/loadFromBlob/buildPortMetadata (public), plus the
previously-private loadFileDirectly/migrateFile/deserializeAndLoad/
processLoadedItems/loadBoundaryElement/loadBoundaryPorts and the
comparePorts/sortPorts/buildPortLabels helpers. Preserves the
qt-cpp-review/fixes-2 exception-safety hardening verbatim: the
qScopeGuard that nulls QNEConnection* before qDeleteAll (both in
loadFileDirectly and deserializeAndLoad), and the non-const
QList<QGraphicsItem*>& drain-via-takeFirst() signature on
processLoadedItems. Gets Q_DECLARE_TR_FUNCTIONS + plain PANDACEPTION on
all 6 throw sites from the start, avoiding the WITH_CONTEXT detour that
would otherwise silently drop these strings from lupdate extraction.

ICRenderer: drawBody/generatePixmap/generatePreviewPixmap, plus the
mascot-logo SVG renderer singleton. No throws, so no tr context needed.

ICSimulation: initialize/update (from initializeSimulation/updateLogic)
plus a new resettle() (from resettleCombinational(), added to master
after the original branch this recreation is modeled on — it shares the
same pushInputsToBoundary/pullOutputsFromBoundary helpers as update()).

IC keeps the constructor/destructor, resetInternalState(), save()/load(),
hover/double-click handlers, displayName(), boundingRect(), paint()
(still touches m_appearance directly for the selection-rect overlay),
refresh(), externalFiles(), loadFromDrop(), isCursorOverPort(), and all
data members. Its public loadFile/loadFromBlob/buildPortMetadata/
initializeSimulation/updateLogic/resettleCombinational become one-line
delegators, keeping their exact names since updateLogic/
initializeSimulation/resettleCombinational are virtual overrides
dispatched polymorphically and buildPortMetadata has an external
non-member caller (Workspace.cpp) that must keep compiling unchanged.

IC.h gains friend class ICLoader/ICRenderer/ICSimulation; the protected
GraphicElement members these reach (pixmap(), portsBoundingRect(),
setMax/MinInputSize/OutputSize(), simUpdateInputsAllowUnknown(),
m_appearance, m_label) are accessible because every collaborator method
takes IC& (not GraphicElement&) as its object-expression type.

Also fixes a latent missing-include gap in
MCP/Server/Handlers/SimulationHandler.cpp exposed by the unity-build
batch reshuffle from the 3 new .cpp files: it constructs a
BewavedDolphin(..., m_mainWindow) where the third parameter is
DolphinHost*, relying on the implicit MainWindow*->DolphinHost* upcast,
but never included MainWindow.h (BaseHandler.h only forward-declares
MainWindow for its pointer member). This only compiled because whichever
other file shared its unity translation unit happened to pull in
MainWindow.h first — the same class of bug fixed for ICPreviewPopup.h in
the ElementPorts commit. Add the missing include directly.
